+++ workspaces/orchestrator/plugins/orchestrator/src/components/WorkflowInstancePage/WorkflowInstancePage.tsx
@@ -75,7 +75,6 @@ import { orchestratorTranslationRef } from '../../translations';
 import { deepSearchObject } from '../../utils/deepSearchObject';
 import { isNonNullable } from '../../utils/TypeGuards';
 import { buildUrl } from '../../utils/UrlUtils';
-import { Trans } from '../Trans';
 import { BaseOrchestratorPage } from '../ui/BaseOrchestratorPage';
 import { InfoDialog } from '../ui/InfoDialog';
 import { WorkflowInstancePageContent } from './WorkflowInstancePageContent';
@@ -391,12 +390,9 @@ export const WorkflowInstancePage = () => {
 
   const combinedError: Error | undefined = error || inputSchemaError;
 
-  const title = (
-    <Trans
-      message="run.pageTitle"
-      params={{ processName: value?.processName }}
-    />
-  ) as unknown as string;
+  const title = t('run.pageTitle', {
+    processName: value?.processName ?? '',
+  });
 
   return (
     <BaseOrchestratorPage title={title}>