+++ workspaces/lightspeed/packages/backend/CHANGELOG.md
@@ -1,5 +1,20 @@
 # backend
 
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies [99f35d5]
+  - @red-hat-developer-hub/backstage-plugin-lightspeed-backend@1.1.0
+
+## 0.0.23
+
+### Patch Changes
+
+- Updated dependencies [234cc73]
+- Updated dependencies [40b80fe]
+  - @red-hat-developer-hub/backstage-plugin-lightspeed-backend@1.0.4
+
 ## 0.0.22
 
 ### Patch Changes