+++ workspaces/lightspeed/plugins/lightspeed-backend/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @red-hat-developer-hub/backstage-plugin-lightspeed-backend
 
+## 1.1.0
+
+### Minor Changes
+
+- 99f35d5: Backstage version bump to v1.45.2
+
+### Patch Changes
+
+- Updated dependencies [99f35d5]
+  - @red-hat-developer-hub/backstage-plugin-lightspeed-common@1.1.0
+
+## 1.0.4
+
+### Patch Changes
+
+- 234cc73: Updated dependency `msw` to `2.12.2`.
+- 40b80fe: Remove "lifecycle" keywords and "supported-versions" in package.json. Change "lifecycle" to active in catalog.yaml
+- Updated dependencies [58e26ba]
+- Updated dependencies [40b80fe]
+  - @red-hat-developer-hub/backstage-plugin-lightspeed-common@1.0.4
+
 ## 1.0.3
 
 ### Patch Changes