+++ workspaces/lightspeed/plugins/lightspeed-common/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @red-hat-developer-hub/backstage-plugin-lightspeed-common
 
+## 1.1.0
+
+### Minor Changes
+
+- 99f35d5: Backstage version bump to v1.45.2
+
+## 1.0.4
+
+### Patch Changes
+
+- 58e26ba: '@red-hat-developer-hub/backstage-plugin-lightspeed' :
+  - Added missing translations for untranslated labels across the Lightspeed chat UI.
+  - Added lightspeed settings kebab menu
+  - Added a new “Pinned” section. The section is conditionally visible only when pinning chats are enabled in Lightspeed settings. It shows an empty state when no chats are pinned.
+  - Added Kebab menu actions to each chat item with actions: Rename, Pin, Unpin
+  - Implemented logic to move chats between Pinned and Recent sections based on the selected action (“Pin” / “Unpin”).
+  - Updated search behavior to show results under both Pinned and Recent groups (when enabled), including appropriate empty search messaging.
+  - The Rename action opens a dialog with an editable text field and confirm/cancel actions.
+  - Replaced the patternfly components with MUI to match the Figma designs [Design 1](https://www.figma.com/design/urwU8VqRvHfbxMqxeVknrv/RHDH-Lightspeed---AI-Assistant?node-id=1523-1670&t=CqnUYsxar0PtjAjZ-0)
+    [Design 2](https://www.figma.com/design/urwU8VqRvHfbxMqxeVknrv/RHDH-Lightspeed---AI-Assistant?node-id=2514-5415&t=PphIEOl5Yzmgv5tI-0)
+
+  '@red-hat-developer-hub/backstage-plugin-lightspeed-common':
+
+  Added the Update api client
+
+- 40b80fe: Remove "lifecycle" keywords and "supported-versions" in package.json. Change "lifecycle" to active in catalog.yaml
+
 ## 1.0.3
 
 ## 1.0.2