+++ workspaces/marketplace/packages/cli/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @red-hat-developer-hub/marketplace-cli
 
+## 0.13.0
+
+### Minor Changes
+
+- bee615a: Backstage version bump to v1.45.2
+
+### Patch Changes
+
+- Updated dependencies [bee615a]
+  - @red-hat-developer-hub/backstage-plugin-marketplace-common@0.13.0
+
 ## 0.12.2
 
 ### Patch Changes