+++ workspaces/orchestrator/plugins/orchestrator/CHANGELOG.md
@@ -1,5 +1,31 @@
 ### Dependencies
 
+## 5.3.0
+
+### Minor Changes
+
+- 782c33f: Removal and updating outdated and unmaintained dependencies
+- 29dfed0: Backstage version bump to v1.45.2
+
+### Patch Changes
+
+- a970cf3: Fix workflow start time displaying `{{ time }}` placeholder instead of actual time
+- 40b80fe: Change "lifecycle" to active in catalog-info.yaml
+- f5f4973: Updated dependency `@openapitools/openapi-generator-cli` to `2.25.0`.
+  Updated dependency `@red-hat-developer-hub/backstage-plugin-theme` to `^0.10.0`.
+- 40b80fe: Remove "support", "lifecycle" keywords and "supported-versions" in package.json. Change "lifecycle" to active in catalog.yaml
+- Updated dependencies [a1671ab]
+- Updated dependencies [40b80fe]
+- Updated dependencies [782c33f]
+- Updated dependencies [f5f4973]
+- Updated dependencies [40b80fe]
+- Updated dependencies [34a36cb]
+- Updated dependencies [6db1430]
+- Updated dependencies [29dfed0]
+  - @red-hat-developer-hub/backstage-plugin-orchestrator-common@3.3.0
+  - @red-hat-developer-hub/backstage-plugin-orchestrator-form-react@2.4.0
+  - @red-hat-developer-hub/backstage-plugin-orchestrator-form-api@2.4.0
+
 ## 5.2.1
 
 ### Patch Changes